ric_auth-dev/requirements/base.txt
Django==3.1.3
psycopg2-binary==2.8.6
djangorestframework==3.12.1
pyyaml==5.3.1
uritemplate==3.0.1
drf-spectacular==0.11.1
django-filter==2.4.0
djangorestframework-simplejwt==4.6.0
PyJWT==2.0.0a1
Pillow==8.0.1
celery==5.0.4
redis==3.5.3
django-redis==4.12.1
sendgrid==6.5.0
